Update Dockerfile and build scripts for Ubuntu 24.04
This commit improves the Docker build process and the Emacs installation script. Ubuntu:24.04 LTS is
now the base image. The source packages configuration has been modified to reflect changes in the
location of system packages and the build is carried out by the ubuntu user to not taint the local
file-system.